requirements.txt
autogen-agentchat==0.4.0.dev12
autogen-core==0.4.0.dev12
autogen-ext[azure,ioenai]==0.4.0.dev12
chainlit
python-dotenv
pydantic==2.10.1
opentelemetry-sdk==1.28.2
